+++ pieslicecell.py
@@ -44,7 +44,7 @@ class PieSliceCell(Cell):
 
     def draw(self, N=2):
         # super sample by 3x
-        n_width, n_height = (self.width*N, self.height*N)
+        n_width, n_height = int(self.width*N), int(self.height*N)
         paper = Image.new('RGBA', (n_width, n_height))
         canvas = ImageDraw.Draw(paper, paper.mode)
