+++ rectcell.py
@@ -54,7 +54,7 @@ class RectCell(Cell):
 
     # return the perceived hue / luminance for now
     def draw(self, N=2):
-        n_width, n_height = self.width*N, self.height*N
+        n_width, n_height = int(self.width*N), int(self.height*N)
         paper = Image.new('RGBA', (n_width, n_height))
         canvas = ImageDraw.Draw(paper)
